# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BanHang Backend API
  version: 2.0.0
  description: Versioned NestJS backend API for auth, CRM, content, catalog, cart, orders, payments, inventory, shipping, notifications, reporting, and store locator flows.
servers:
  - url: http://localhost:4000/api/v1
...
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: List orders for current user or all orders for privileged users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Order list
    post:
      tags: [Orders]
      summary: Alias of checkout endpoint
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '201':
          description: Order created
  /orders/checkout:
    post:
      tags: [Orders]
      summary: Checkout reservation into order for customer or guest session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '201':
          description: Order created
  /orders/reservations/current:
    get:
      tags: [Orders]
      summary: Get current active reservation
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Reservation summary or null
  /orders/reservations:
    post:
      tags: [Orders]
      summary: Reserve stock from current cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Reservation created
  /orders/reservations/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel reservation and release stock
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reservation canceled
  /orders/reservations/release-expired:
    post:
      tags: [Orders]
      summary: Release expired reservations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Expired reservations released
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get order detail
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order detail
  /orders/{id}/notes:
    get:
      tags: [Orders]
      summary: List order notes with internal notes hidden from customers
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order notes
    post:
      tags: [Orders]
      summary: Add internal or customer-facing note to an order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderNoteRequest'
      responses:
        '201':
          description: Order note created
  /orders/{id}/invoice:
    get:
      tags: [Orders]
      summary: Get invoice payload
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice data
  /orders/{id}/tracking:
    get:
      tags: [Orders]
      summary: Get shipment tracking summary
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tracking data
  /orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order canceled
  /orders/{id}/return:
    post:
      tags: [Orders]
      summary: Request order return
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Return requested
  /orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Advance order status in managed flow
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        '200':
          description: Order updated
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List current user notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notification list
  /notifications/{id}/click:
    patch:
      tags: [Notifications]
      summary: Track notification click
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification click tracked
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Mark a notification as read
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification updated
  /notifications/read/all:
    patch:
      tags: [Notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notifications updated
  /notifications/preferences/current:
    get:
      tags: [Notifications]
      summary: Get current user notification preferences
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notification preferences
    patch:
      tags: [Notifications]
      summary: Update current user notification preferences
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotificationPreferencesRequest'
      responses:
        '200':
          description: Notification preferences updated
  /notifications/preferences/unsubscribe:
    post:
      tags: [Notifications]
      summary: Unsubscribe current user from a notification scope
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnsubscribeNotificationsRequest'
      responses:
        '200':
          description: Notification preferences updated
  /notifications/push/subscriptions:
    get:
      tags: [Notifications]
      summary: List current user Web Push subscriptions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Push subscription list
    post:
      tags: [Notifications]
      summary: Create or refresh current user Web Push subscription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePushSubscriptionRequest'
      responses:
        '201':
          description: Push subscription saved
  /notifications/push/subscriptions/{id}:
    delete:
      tags: [Notifications]
      summary: Remove current user Web Push subscription
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Push subscription removed
  /notifications/templates/list:
    get:
      tags: [Notifications]
      summary: List notification templates
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: channel
          required: false
          schema:
            type: string
            enum: [in_app, email, sms, push]
      responses:
        '200':
          description: Notification template list
  /notifications/templates:
    post:
      tags: [Notifications]
      summary: Create notification template
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationTemplateRequest'
      responses:
        '200':
          description: Notification template created
  /notifications/templates/{id}:
    patch:
      tags: [Notifications]
      summary: Update notification template
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotificationTemplateRequest'
      responses:
        '200':
          description: Notification template updated
  /notifications/templates/{id}/preview:
    post:
      tags: [Notifications]
      summary: Preview rendered notification template
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreviewNotificationTemplateRequest'
      responses:
        '200':
          description: Notification template preview
  /notifications/batch:
    post:
      tags: [Notifications]
      summary: Create notification batch or scheduled campaign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationBatchRequest'
      responses:
        '200':
          description: Notification batch created
  /notifications/dispatch:
    post:
      tags: [Notifications]
      summary: Dispatch due scheduled notifications
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchScheduledNotificationsRequest'
      responses:
        '200':
          description: Scheduled notifications dispatched
  /notifications/dispatch/abandoned-cart:
    post:
      tags: [Notifications]
      summary: Dispatch abandoned cart reminders for idle server-side carts
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchAbandonedCartRemindersRequest'
      responses:
        '200':
          description: Abandoned cart reminders dispatched
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    CheckoutRequest:
      type: object
      required: [reservationId, paymentMethod, shippingMethod]
      properties:
        reservationId:
          type: string
        addressId:
          type: string
        address:
          type: object
          required: [receiverName, phone, district, addressLine, country]
          properties:
            receiverName:
              type: string
            phone:
              type: string
            province:
              type: string
            district:
              type: string
            ward:
              type: string
            addressLine:
              type: string
            country:
              type: string
        saveAddress:
          type: boolean
        paymentMethod:
          type: string
          enum: [cod, vnpay, momo, zalopay, stripe, paypal, bank_transfer]
        shippingMethod:
          type: string
          enum: [standard, express, same_day, pickup]
        notes:
          type: string
    CreateOrderNoteRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 3
          maxLength: 2000
        visibility:
          type: string
          enum: [internal, customer]
    UpdateOrderStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [confirmed, shipping, completed]
...
    UpdateNotificationPreferencesRequest:
      type: object
      properties:
        orderInApp:
          type: boolean
        orderEmail:
          type: boolean
        promotionInApp:
          type: boolean
        promotionEmail:
          type: boolean
        securityInApp:
          type: boolean
        securityEmail:
          type: boolean
        systemInApp:
          type: boolean
        systemEmail:
          type: boolean
        pushEnabled:
          type: boolean
        smsEnabled:
          type: boolean
        marketingOptIn:
          type: boolean
    UnsubscribeNotificationsRequest:
      type: object
      properties:
        type:
          type: string
          enum: [order, system, promotion, security]
        channel:
          type: string
          enum: [in_app, email, sms, push]
    CreatePushSubscriptionRequest:
      type: object
      required: [endpoint, p256dh, auth]
      properties:
        endpoint:
          type: string
          format: uri
        p256dh:
          type: string
          minLength: 8
          maxLength: 512
        auth:
          type: string
          minLength: 8
          maxLength: 256
        userAgent:
          type: string
          maxLength: 255
    CreateNotificationTemplateRequest:
      type: object
      required: [key, channel, titleTemplate, contentTemplate]
      properties:
        key:
          type: string
        channel:
          type: string
          enum: [in_app, email, sms, push]
        subjectTemplate:
          type: string
        titleTemplate:
          type: string
        contentTemplate:
          type: string
        isActive:
          type: boolean
    UpdateNotificationTemplateRequest:
      type: object
      properties:
        key:
          type: string
        channel:
          type: string
          enum: [in_app, email, sms, push]
        subjectTemplate:
          type: string
        titleTemplate:
          type: string
        contentTemplate:
          type: string
        isActive:
          type: boolean
    PreviewNotificationTemplateRequest:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
    CreateNotificationBatchRequest:
      type: object
      required: [userIds, type, channel]
      properties:
        userIds:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [order, system, promotion, security]
        channel:
          type: string
          enum: [in_app, email, sms, push]
        templateId:
          type: string
        title:
          type: string
        content:
          type: string
        campaignKey:
          type: string
        scheduledFor:
          type: string
          format: date-time
        data:
          type: object
          additionalProperties: true
    DispatchScheduledNotificationsRequest:
      type: object
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 500
    DispatchAbandonedCartRemindersRequest:
      type: object
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 500
        idleMinutes:
          type: integer
          minimum: 15
          maximum: 43200
        channel:
          type: string
          enum: [in_app, email]
...